+++ lib/safe-pg-migrations/base.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 require 'safe-pg-migrations/configuration'
+require 'safe-pg-migrations/helpers/satisfied_helper'
+require 'safe-pg-migrations/helpers/add_check_constraint_helper'
 require 'safe-pg-migrations/plugins/verbose_sql_logger'
 require 'safe-pg-migrations/plugins/blocking_activity_logger'
 require 'safe-pg-migrations/plugins/statement_insurer'
 require 'safe-pg-migrations/plugins/statement_retrier'
 require 'safe-pg-migrations/plugins/idempotent_statements'
 require 'safe-pg-migrations/plugins/useless_statements_logger'
-require 'safe-pg-migrations/polyfills/satisfied_helper'
 require 'safe-pg-migrations/polyfills/index_definition_polyfill'
 require 'safe-pg-migrations/polyfills/verbose_query_logs_polyfill'
-require 'safe-pg-migrations/polyfills/add_check_constraint_polyfill'
 
 module SafePgMigrations
   # Order matters: the bottom-most plugin will have precedence
@@ -21,7 +21,6 @@ module SafePgMigrations
     StatementInsurer,
     UselessStatementsLogger,
     Polyfills::IndexDefinitionPolyfill,
-    Polyfills::AddCheckConstraintPolyfill,
   ].freeze
 
   class << self
